Move utils out of /api
