+++ server/src/api/modules/deltas/services/FindGroupDeltasService.ts
@@ -1,11 +1,11 @@
-import { omit } from '../../../util/objects';
-import { Metric, Period, parsePeriodExpression } from '../../../../utils';
 import prisma, { Player, Snapshot } from '../../../../prisma';
+import { Metric, Period, parsePeriodExpression } from '../../../../utils';
+import { omit } from '../../../../utils/omit.util';
 import { BadRequestError, NotFoundError } from '../../../errors';
-import { calculateMetricDelta } from '../delta.utils';
-import { DeltaGroupLeaderboardEntry } from '../delta.types';
-import { findGroupSnapshots } from '../../snapshots/services/FindGroupSnapshotsService';
 import { PaginationOptions } from '../../../util/validation';
+import { findGroupSnapshots } from '../../snapshots/services/FindGroupSnapshotsService';
+import { DeltaGroupLeaderboardEntry } from '../delta.types';
+import { calculateMetricDelta } from '../delta.utils';
 
 async function findGroupDeltas(
   groupId: number,
