+++ server/src/api/modules/groups/services/CreateGroupService.ts
@@ -2,9 +2,9 @@ import { isErrored } from '@attio/fetchable';
 import prisma from '../../../../prisma';
 import * as cryptService from '../../../../services/crypt.service';
 import { GroupRole, PlayerAnnotationType, PRIVELEGED_GROUP_ROLES } from '../../../../utils';
+import { omit } from '../../../../utils/omit.util';
 import { BadRequestError, ForbiddenError } from '../../../errors';
 import { eventEmitter, EventType } from '../../../events';
-import { omit } from '../../../util/objects';
 import { isValidUsername, sanitize, standardize } from '../../players/player.utils';
 import { findOrCreatePlayers } from '../../players/services/FindOrCreatePlayersService';
 import { GroupDetails } from '../group.types';
