+++ server/src/api/modules/groups/services/SearchGroupsService.ts
@@ -1,5 +1,5 @@
 import prisma from '../../../../prisma';
-import { omit } from '../../../util/objects';
+import { omit } from '../../../../utils/omit.util';
 import { PaginationOptions } from '../../../util/validation';
 import { GroupListItem } from '../group.types';
