+++ server/src/jobs/handlers/sync-patrons.job.ts
@@ -1,6 +1,5 @@
 import { isErrored } from '@attio/fetchable';
 import { eventEmitter, EventType } from '../../api/events';
-import { omit } from '../../api/util/objects';
 import prisma, { Patron } from '../../prisma';
 import { sendDiscordWebhook } from '../../services/discord.service';
 import {
@@ -8,6 +7,7 @@ import {
   STATIC_PATRON_GROUP_IDS,
   STATIC_PATRON_PLAYER_IDS
 } from '../../services/patreon.service';
+import { omit } from '../../utils/omit.util';
 import { Job } from '../job.class';
 
 export class SyncPatronsJob extends Job<unknown> {
