+++ server/src/utils/format-date.util.ts
@@ -1,0 +1,5 @@
+import dayjs from 'dayjs';
+
+export function formatDate(date: Date, mask = 'MM-DD-YYYY HH:mm') {
+  return dayjs(date).format(mask);
+}
