+++ server/src/utils/is-valid-date.util.ts
@@ -1,0 +1,5 @@
+import dayjs from 'dayjs';
+
+export function isValidDate(dateString: string) {
+  return dateString && dayjs(dateString).isValid();
+}
