+++ server/src/utils/map-values.util.ts
@@ -1,16 +1,3 @@
-export function omit<T extends object, K extends keyof T>(
-  object: T,
-  ...fields: K[]
-): Pick<T, Exclude<keyof T, K>> {
-  const clone = { ...object };
-
-  fields.forEach(f => {
-    delete clone[f as keyof object];
-  });
-
-  return clone;
-}
-
 export function mapValues<T extends object, TResult>(
   obj: T,
   callback: (value: T[keyof T], key: keyof T, collection: T) => TResult
