+++ server/src/utils/metrics.ts
@@ -1,5 +1,5 @@
-import { Skill, Boss, Activity, ComputedMetric, Metric } from '../prisma/enum-adapter';
-import { mapValues } from '../api/util/objects';
+import { Activity, Boss, ComputedMetric, Metric, Skill } from '../prisma/enum-adapter';
+import { mapValues } from './map-values.util';
 import { MapOf } from './types';
 
 enum MetricType {
@@ -287,38 +287,38 @@ function getParentEfficiencyMetric(metric: Metric) {
 }
 
 export {
+  ACTIVITIES,
+  Activity,
+  Boss,
+  BOSSES,
+  COMBAT_SKILLS,
+  COMPUTED_METRICS,
+  ComputedMetric,
+  F2P_BOSSES,
+  // Functions
+  findMetric,
+  getMetricMeasure,
+  getMetricName,
+  getMetricRankKey,
+  getMetricValueKey,
+  getMinimumValue,
+  getParentEfficiencyMetric,
+  isActivity,
+  isBoss,
+  isComputedMetric,
+  isMetric,
+  isSkill,
+  MEMBER_SKILLS,
   // Enums
   Metric,
-  Skill,
-  Boss,
-  Activity,
-  ComputedMetric,
-  MetricType,
   MetricMeasure,
   // Maps
   MetricProps,
-  // Lists
-  SKILLS,
-  ACTIVITIES,
-  BOSSES,
-  COMPUTED_METRICS,
   METRICS,
-  F2P_BOSSES,
-  REAL_SKILLS,
-  MEMBER_SKILLS,
-  COMBAT_SKILLS,
+  MetricType,
   REAL_METRICS,
-  // Functions
-  findMetric,
-  getMetricRankKey,
-  getMetricValueKey,
-  getMetricMeasure,
-  getMetricName,
-  getMinimumValue,
-  getParentEfficiencyMetric,
-  isMetric,
-  isSkill,
-  isActivity,
-  isBoss,
-  isComputedMetric
+  REAL_SKILLS,
+  Skill,
+  // Lists
+  SKILLS
 };
